requirements.txt
# Dependencies for Plant Disease Detection App
streamlit>=1.28.0
tensorflow>=2.13.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
opencv-python>=4.8.0
h5py>=3.9.0
